Create idw.py
This script performs spatial interpolation on precipitation data using the Inverse Distance Weighting (IDW) method.

The script follows these steps:
1. Sets up the environment and checks out the Spatial Analyst extension.
2. Retrieves unique dates from the 'Date' field of the input point features.
3. Sets the extent to the clip boundary.
4. Loops over each unique date and performs the following steps for each date:
    a. Creates a feature layer for the current date.
    b. Executes IDW on the feature layer to interpolate the precipitation data.
    c. Clips the output of the IDW to the specified boundary.
    d. Deletes the temporary feature layer.
5. Prints the end time after all dates have been processed.

The input point features should have a 'Date' field and a field for the values to be interpolated (specified by 'zField'). 
The output is a set of rasters named "IDW" followed by the date, clipped to the specified boundary.

Note: This script requires the ArcGIS Spatial Analyst extension.